Calculate prospective and retrospective P-values for GEE or GLMM model

This function tests a SNPs for a given SNP set for a given lbrat estimated null model.

lbrat_test(lbrat.est, G, impute.method = "fixed", GRM = NULL)

Arguments

lbrat.est
The output of function "lbrat_est()"

G
The genotype matrix, an m*q matrix where m is the number of subjects and q is the total number genetic variants.

impute.method
choose the iputation method when there is missing genotype. Optional options are: 'random', 'fixed' or 'bestguess'.

GRM
takes m-by-m genetic correlation matrix or kinship matrix.

Value

This function returns a dataframe. The row name is the SNP ID, the first column is the prospective score statistics, the second colum is the retrospective score statistics, the third column is the prospective pvalue and the forth column is the restrospective pvalue